Sample app now shows an example of 'install_files' usage

+++ ansible/provisioning/apps/sample_app/install.sh
@@ -2,4 +2,16 @@
 ## Optional file if you need to install some packages on the containers
 
 apt-get -y update
-apt-get -y install stress
+
+if [ -d /opt/{{ app_dir | basename }}/{{ install_files | basename  }} ] && [ -f /opt/{{ app_dir | basename }}/{{ install_files | basename  }}/package_config.sh ];
+then
+    # Install other package
+    source /opt/{{ app_dir | basename }}/{{ install_files | basename }}/package_config.sh
+    if [ $INSTALL_STRESS_NG = 1 ];
+    then
+        apt-get -y install stress-ng
+    fi
+else
+    # Install stress
+    apt-get -y install stress
+fi
